pg_parquet_demo/pyproject.toml
[project]
name = "pg-parquet-client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.9",
]
